# contracts/openapi/case-engine-api.yaml
openapi: "3.0.3"
info:
  title: Case Engine API
  description: |
    REST API for the LL-task-tracker case engine. This is the canonical contract
    for all consumers (including ll-corporate). The case engine is the system of
    record for lifecycle state, workflow transitions, permissions, and operational control.

    Domain authority: ll-task-tracker is the canonical Matter domain owner.
    See contracts/domain/AUTHORITY-MAP.md for full cross-domain authority rules.
  version: "1.1.0"
  contact:
    name: Levine Law Platform Team
servers:
  - url: https://matters.levinellp.ca
    description: Production
  - url: http://localhost:8081
    description: Local development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Keycloak-issued JWT. Obtain from the Keycloak realm token endpoint.
  schemas:
...
    AdminState:
      type: string
      description: Operational state of a matter within its lifecycle stage.
      enum:
        - "Intake Review"
        - "Awaiting Engagement"
        - "Ready to Open"
        - "Ready for Lawyer"
        - "Waiting on Client"
        - "Opened"
        - "Active"
        - "Maintenance Lawyer Review"
        - "Maintenance Client Wait"
        - "Waiting on External"
        - "Closing Review"
        - "Closed"
        - "Archived"
...
    AdminHealth:
      type: string
      description: Health signal for a matter based on automated control evaluation.
      enum:
        - "Green"
        - "Amber"
        - "Red"
...
    NextActionOwnerType:
      type: string
      description: Who owns the next action on a matter.
      enum:
        - Admin
        - Lawyer
        - Client
        - External
        - System
...
    CaseStatus:
      type: string
      description: Process-level status of a case instance.
      enum:
        - open
        - closed

    # ── Component Objects ──────────────────────────────────────────────────────
...
  responses:
    Unauthorized:
      description: Missing or invalid Bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
paths:
...
  /case:
    get:
      operationId: listCases
      summary: List case instances
      description: |
        Returns a cursor-paginated list of case instances matching the given filters.
        Pagination uses opaque cursors. Do not construct or parse cursor values.
      tags: [Case Instance]
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/CaseStatus"
        - name: caseDefinitionId
          in: query
          schema:
            type: string
        - name: adminState
          in: query
          schema:
            $ref: "#/components/schemas/AdminState"
        - name: adminHealth
          in: query
          schema:
            $ref: "#/components/schemas/AdminHealth"
        - name: nextActionOwnerType
          in: query
          schema:
            $ref: "#/components/schemas/NextActionOwnerType"
        - name: queueId
          in: query
          schema:
            type: string
        - name: adminOwnerId
          in: query
          schema:
            type: string
        - name: healthReasonCode
          in: query
          schema:
            $ref: "#/components/schemas/AdminHealthReasonCode"
        - name: malformedCase
          in: query
          schema:
            type: string
        - name: exceptionOnly
          in: query
          schema:
            type: string
        - name: before
          in: query
          description: Cursor for the previous page (from paging.cursors.before).
          schema:
            type: string
        - name: after
          in: query
          description: Cursor for the next page (from paging.cursors.after).
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            description: "Field to sort by, e.g. lastStateChangedAt or nextActionDueAt."
        - name: limit
          in: query
          schema:
            type: string
            description: Maximum number of results to return.
      responses:
        "200":
          description: Paginated list of matching case instances.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseInstancePage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

    post:
      operationId: createCase
      summary: Create a new case instance
      tags: [Case Instance]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaseInstance"
      responses:
        "200":
          description: The newly created case instance.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseInstance"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /case/{businessKey}:
    get:
      operationId: getCase
      summary: Get a case instance by business key
      tags: [Case Instance]
      parameters:
        - name: businessKey
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The case instance.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseInstance"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      operationId: patchCase
      summary: Partially update non-lifecycle fields on a case
      description: |
        JSON merge patch (RFC 7396). Admin lifecycle fields (adminState, adminHealth,
        responsibleLawyerId, nextAction*, waiting*, etc.) are blocked — use the
        transition endpoint instead.
      tags: [Case Instance]
      parameters:
        - name: businessKey
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: "#/components/schemas/CaseInstance"
      responses:
        "204":
          description: Patch applied.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      operationId: deleteCase
      summary: Delete a case instance
      tags: [Case Instance]
      parameters:
        - name: businessKey
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Case deleted.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /case/{businessKey}/transition/{transitionName}:
    post:
      operationId: transitionCase
      summary: Trigger a named lifecycle transition on a case
      description: |
        Applies one of the defined AdminTransition values to move the case to
        the next lifecycle state. The case engine validates whether the transition
        is allowed given the current state and role. Consumers MUST NOT infer
        allowed transitions from the current adminState — always defer to the
        backend for eligibility.
      tags: [Case Instance]
      parameters:
        - name: businessKey
          in: path
          required: true
          schema:
            type: string
        - name: transitionName
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/AdminTransition"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminTransitionRequest"
      responses:
        "200":
          description: The updated case instance after transition.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseInstance"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /case/{businessKey}/accounts:
    get:
      operationId: getCaseAccounts
      summary: Get the accounts lifecycle shell for a case
      description: |
        Initializes the accounts lifecycle shell when missing and returns
        the current case snapshot. This endpoint does not perform readiness,
        health, activation, or cross-lifecycle blocking.
      tags: [Case Instance]
      parameters:
        - name: businessKey
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Case instance with accounts shell fields.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseInstance"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /case/accounts/work:
    get:
      operationId: listAccountsWork
      summary: List outstanding and upcoming accounts work
      description: |
        Returns matters with accounts work signals. This is a filtered work view
        over matter/account fields and does not define lifecycle stage or state.
      tags: [Case Instance]
      parameters:
        - name: accountsReadinessStatus
          in: query
          schema:
            $ref: "#/components/schemas/AccountsReadinessStatus"
        - name: accountsQueueId
          in: query
          schema:
            type: string
        - name: accountsNextActionOwnerType
          in: query
          schema:
            $ref: "#/components/schemas/AccountsWorkOwnerType"
        - name: accountsNextActionDueBefore
          in: query
          schema:
            type: string
            format: date
        - name: accountsWorkBlocked
          in: query
          schema:
            type: string
        - name: accountsReadinessReasonCode
          in: query
          schema:
            $ref: "#/components/schemas/AccountsHealthReasonCode"
        - name: matterType
          in: query
          schema:
            $ref: "#/components/schemas/MatterType"
        - name: adminOwnerId
          in: query
          schema:
            type: string
        - name: responsibleLawyerId
          in: query
          schema:
            type: string
        - name: before
          in: query
          schema:
            type: string
        - name: after
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated accounts work matters.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseInstancePage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /case/accounts/work/summary:
    get:
      operationId: summarizeAccountsWork
      summary: Summarize outstanding and upcoming accounts work
      description: Returns counts by queue, owner, readiness status, blocked, and due bucket.
      tags: [Case Instance]
      parameters:
        - name: accountsReadinessStatus
          in: query
          schema:
            $ref: "#/components/schemas/AccountsReadinessStatus"
        - name: accountsQueueId
          in: query
          schema:
            type: string
        - name: accountsNextActionOwnerType
          in: query
          schema:
            $ref: "#/components/schemas/AccountsWorkOwnerType"
        - name: accountsNextActionDueBefore
          in: query
          schema:
            type: string
            format: date
        - name: accountsWorkBlocked
          in: query
          schema:
            type: string
        - name: accountsReadinessReasonCode
          in: query
          schema:
            $ref: "#/components/schemas/AccountsHealthReasonCode"
        - name: matterType
          in: query
          schema:
            $ref: "#/components/schemas/MatterType"
        - name: adminOwnerId
          in: query
          schema:
            type: string
        - name: responsibleLawyerId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Accounts work summary.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountsWorkSummary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /case/{businessKey}/accounts/history:
    get:
      operationId: getCaseAccountsHistory
      summary: Get accounts lifecycle event history for a case
      description: Returns persisted accounts lifecycle events.
      tags: [Case Instance]
      parameters:
        - name: businessKey
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Accounts lifecycle event history.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccountsEvent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /case/{businessKey}/accounts/readiness:
    get:
      operationId: getCaseAccountsReadiness
      summary: Get accounts readiness for administrative opening
      description: Returns the last persisted accounts readiness evaluation, or NOT_EVALUATED when no evaluation has run.
      tags: [Case Instance]
      parameters:
        - name: businessKey
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Accounts readiness evaluation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountsReadinessEvaluation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /case/{businessKey}/accounts/evaluate-readiness:
    post:
      operationId: evaluateCaseAccountsReadiness
      summary: Evaluate accounts readiness for administrative opening
      description: |
        Evaluates accounts setup through the resolved matter-type policy and persists
        readiness status, structured reason codes, timestamp, and summary.
      tags: [Case Instance]
      parameters:
        - name: businessKey
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Fresh accounts readiness evaluation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountsReadinessEvaluation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /case/{businessKey}/document:
    post:
      operationId: saveDocument
      summary: Attach a document to a case
      tags: [Case Instance]
      parameters:
        - name: businessKey
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaseDocument"
      responses:
        "204":
          description: Document attached.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...